requirements.txt
Flask==2.3.3
Pillow==10.1.0
pycryptodome==3.19.0
piexif==1.1.3